[ark-app] refactor: move ark level files to app
