+++ ArkKit/ArkBlueprint.swift
@@ -102,15 +102,3 @@ struct ArkBlueprint<ExternalResources: ArkExternalResources> {
         return newSelf
     }
 }
-
-struct ArkNetworkPlayableInfo {
-    let roomName: String
-    let numberOfPlayers: Int
-    let role: ArkPeerRole?
-
-    init(roomName: String, numberOfPlayers: Int, role: ArkPeerRole? = nil) {
-        self.roomName = roomName
-        self.numberOfPlayers = numberOfPlayers
-        self.role = role
-    }
-}
