+++ ArkKit/app/utils/network/ArkNetworkPlayableInfo.swift
@@ -1,0 +1,11 @@
+struct ArkNetworkPlayableInfo {
+    let roomName: String
+    let numberOfPlayers: Int
+    let role: ArkPeerRole?
+
+    init(roomName: String, numberOfPlayers: Int, role: ArkPeerRole? = nil) {
+        self.roomName = roomName
+        self.numberOfPlayers = numberOfPlayers
+        self.role = role
+    }
+}
